requirements.txt
spiceypy
pandas
numpy
rich
rich-click
PyYaml 
geopandas
planetary-coverage